Fix pointer signedness warnings
It's a bit ugly that GLib string functions take signed chars and
SQLite'3 string functions return unsigned chars. We could disable
that warning instead but maybe it's useful in some other places.

+++ src/libtracker-extract/tracker-utils.c
@@ -822,7 +822,7 @@ tracker_date_guess (const gchar *date_string)
 
 static gint
 my_igetdelim (gchar  **linebuf,
-              guint   *linebufsz,
+              gsize   *linebufsz,
               gint     delimiter,
               FILE    *file)
 {
